requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
django-filter==22.1
Django==4.0.6
djangorestframework==3.13.1
drf-yasg==1.21.3
idna==3.3
importlib-metadata==4.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
packaging==21.3
pip==22.1.2
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
ruamel.yaml.clib==0.2.6
ruamel.yaml==0.17.21
setuptools==62.6.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.10
wheel==0.37.1
zipp==3.8.1
psycopg2>=2.8.6,<2.9
